pyproject.toml
[project]
name = "songon-tour"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-filer>=3.4.1",
    "django-fsm<3",
    "djangocms-alias>=3.0.2",
    "djangocms-frontend>=2.0.0a1",
    "djangocms-simple-admin-style>=2.0.0",
    "djangocms-text>=0.9.3",
    "djangocms-versioning>=2.5.0",
]
